css/css-shadow-parts/chaining-negative.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Chaining Negative</title>
    <meta href="fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="http://www.w3.org/TR/css-variables-1/#syntax" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>#c-e-outer::part(c-e-part)::part(partp) { color: #00ff00; }</style>
    <script type="text/javascript">installCustomElement("custom-element-inner", "custom-element-inner-template");</script>
    <template id="custom-element-inner-template">
      <style>span { color: #ff0000; }</style>
      <span id="part" part="partp">This text</span>
    </template>
    <script type="text/javascript">installCustomElement("custom-element-outer", "custom-element-outer-template");</script>
    <template id="custom-element-outer-template">
      The following text should be red:
      <custom-element-inner id="c-e-inner" part="c-e-part"></custom-element-inner>
    </template>
    The following text should be red:
    <custom-element-outer id="c-e-outer"></custom-element-outer>
    <script type="text/javascript">
      "use strict";
      test(function() {
        var el = getElementByShadowIds(document, ["c-e-outer", "c-e-inner", "part"]);
        assert_equals(window.getComputedStyle(el).color, "rgb(255, 0, 0)");
      }, "::part cannot be chained to reach elements in the inner host");
    </script>
  </body>
</html>
